Extract stac search pagination into a hook and refactor
